upgrade react and lodash
Archive the old libraries and pull in some new ones. I'm using
package.json to manage their versions and babel as a glorified `cp`
mechanism. I'm trying hard to avoid webpack because it's complicated.
I'm still using the production versions because my workflow for
development is to build on my laptop and serve over wifi to my desktop,
and it does take non-trivial amounts of time to transfer everything (and
I typically disable caching so the css file updates).