add "uptime" stat
This is kind of buggy. It's maybe very buggy. The data is straight from
ACT, but ACT sometimes says the encounter lasted 2:52 but someone had
3:04 of "duration", whatever that means. So it's not perfect, and maybe
I'll remove it. It was at least a good excuse to refactor the formatting
of the rows.

I suspect that something naive is happening here. Like I see fights
where someone died 3 times but their uptime is pretty high. It could
just mean that while they were dead, their dots were ticking and that
counted as "uptime". Will keep this in dev for a bit to test it out.