meters: rewrite uptime tracking to use log lines
Now that I sort of know how this works, this is a better source of truth
when it comes to GCDs. I also realized that casts > GCD (e.g., flare)
need to be credited higher -- and really all casts in general need to be
credited at cast start, not cast finish. So this extra data worked out.

In some light testing, it appears to be about right, though it seems to
interact a bit weirdly with ACT when it trims downtime. For example, in
one fight I was 12s behind the encounter clock, but by the end I was
only 8s. I definitely didn't have perfect uptime either, but there was
also fight downtime, but based on what I've seen from ACT so far, the
time tracking is kind of bogus and I almost trust this brand new code
more.

I expanded the README with some new notes.