timers: refactor to simplify data structure
Instead of a triple map, tuple up the `(action, source)` pair together
to form a key.

Except after I committed this I forgot that structural equality doesn't
exist in JavaScript, so I have to store instances of the `TimerKey` in a
global table. This should be fine leakage-wise, but it's certainly more
hassle than you'd like.