meters: more encounter duration improvements
One thing I noticed is that sometimes you get data updates where the
time isn't supposed to advance, but the log data has been ticking away
in the meantime. There's a chance this is simply the final update
marking the encounter as inactive, but for now I went with a more
conservative approach of only updating log data when the ACT timer
ticks.

This came with some refactoring, and at this point I feel like
TypeScript would be fairly useful. I guess it was only a matter of time.