meters: fix player name saving
I thought that the 02 event wasn't firing, but actually I had a guard
that only updated localStorage when the name changed, *and* I defaulted
the name to `options.you`, which meant that it never worked for me. Once
I realized this, I also learned that the storage key was missing.