meters: refactor to parse early
This has been a pain point for a long time. The raw ACT updates are kind
of awful. They're all strings, so we had parsing bits strewn about the
codebase. Now we have separate types for the update and the new offical
`Encounter` type. I added some more structure and picked better names to
make it easier to navigate too.

I even added an upgrade funtion for loading history because it was a
cheap way to make sure I was doing it right. I did run into the perils
of serializing `Date`s as JSON, but I'm fine with some temporary
ugliness in `upgrade`.