more typescript adoption and code-sharing
I had to change `moduleResolution` because I do `import` things from
`node_modules`, which webpack handles for me. Previously it only worked
because the things I was importing were in @types/ packages.

As part of this, I also had to get rid of no-unused-vars in favor of
typescript's linting options.