meters: fix trivial timing edge-case
This came up in development but wouldn't really affect anyone in
reality. I noticed that casting a single DoT on a target dummy would
cause the encounter to start but the time to be stuck at 0:00. At first
I thought it was because our `rollingLogData` gets reset but only via a
`setState`, so I did the correct thing of using the reference. But then
I noticed that the duration was 22 -- the value from ACT itself. We were
simply putting a number representing seconds into a value that was
supposed to be milliseconds.

That said, we shouldn't actually need that. We always have log data at
that point, so the only time that fires is if the duration is 0. The
real issue was that no actions were taken to update the log data. We now
handle that by advancing the clock on any log line with a slight limiter
on update frequency.